// lib/validators/signup-schema.ts
// lib/validators/signup-schema.ts

import { z } from "zod";

export const signupSchema = z.object({
  name: z.string().min(1, "Name is required"),
  email: z.string().email("Invalid email address"),
  password: z.string().min(6, "Password must be at least 6 characters"),
});

export type SignUpSchema = z.infer<typeof signupSchema>;